+++ polywit/_typing/__init__.py
@@ -1,0 +1,6 @@
+from polywit._typing.aliases import Assumption, Position
+
+__all__ = [
+    Assumption,
+    Position
+]
