--- polywit/types/__init__.py
@@ -1,6 +1,0 @@
-from polywit.types.aliases import Assumption, Position
-
-__all__ = [
-    Assumption,
-    Position
-]
