+++ tests/base/test_file_processors.py
@@ -6,11 +6,12 @@ from polywit.base import WitnessProcessor
 
 class TestWitnessProcessor(TestCase):
 
-    def setUp(self):
-        self.tmp_dir = tempfile.mkdtemp()
-        self.witness_processor = WitnessProcessor(self.tmp_dir)
-
-    def test_get_file_name_from_path(self):
+    def test_get_file_name_from_valid_path(self):
         path = 'home/benchmarks/SomeFileName.java'
-        file_name = self.witness_processor._get_file_name_from_path(path)
-        self.assertEqual('SomeFileName1', file_name)
+        file_name = WitnessProcessor._get_file_name_from_path(path)
+        self.assertEqual('SomeFileName', file_name)
+
+    def test_get_file_name_from_valid_path_without_extension(self):
+        path = 'home/benchmarks/SomeFileName'
+        file_name = WitnessProcessor._get_file_name_from_path(path)
+        self.assertEqual('SomeFileName', file_name)
